requirements.txt
streamlit==1.49.1
python-dotenv==1.1.1
nltk==3.9.1
unstructured==0.18.14
unstructured[pdf]==0.18.14
langchain-community==0.3.29
langchain-huggingface==0.3.1
langchain-chroma==0.2.5
langchain-text-splitters==0.3.11
langchain-groq==0.3.7
transformers==4.56.0
sentence-transformers==5.1.0
youtube-search-python==1.6.6
httpx==0.27.0
